signup code updates
From testing, I have discovered that signup is broken. So signup via form will not work for now. It seems to encounter a 401 unauthorized error. I have left some debugging ```console.logs``` to be able to discover the error code that is causing this and hopefully we can solve it. I got rate limited when testing signup. An issue for this will be opened in a bit.