Update ProfilePage.tsx
Implemented a test function that changes the Profile Page based on the context. For example if a user is viewing themselves it displays "Edit Profile" otherwise it displays "Follow" or "Following" (that changes as the button itself is clicked). The logic is NOT implemented, it is just a placeholder to view the different versions after manually setting the context in the form of a state variable.