Rendering create-tweet on all auth states(Please consult description)
When the user is not authenticated, text which is un-styled shows which will definitely not pass during demo. This serves to render the create tweet component which already has a button that tells the user to login if they wish to post. They won't be able to post but additionally, this view provides a standardized view of what the homepage would look like for the user once they do login